app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.9.3",
    "aiohttp-cors>=0.7.0",
    "azure-identity==1.18.0",
    "azure-search-documents==11.6.0b4",
    "azure-storage-blob==12.23.1",
    "google-genai>=1.37.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "openai==1.58.1",
    "pillow>=11.3.0",
    "python-dotenv==1.0.1",
    "rich>=14.1.0",
]
